+++ mlwiz/build_dataset.py
@@ -5,7 +5,8 @@ import sys
 import yaml
 
 from mlwiz.data.util import preprocess_data
-from mlwiz.static import CONFIG_FILE_CLI_ARGUMENT, CONFIG_FILE
+from mlwiz.static import CONFIG_FILE_CLI_ARGUMENT, CONFIG_FILE, \
+    SKIP_SPLITS_CHECK_CLI_ARGUMENT, SKIP_SPLITS_CHECK
 
 
 def get_args_dict() -> dict:
@@ -23,6 +24,13 @@ def get_args_dict() -> dict:
         dest=CONFIG_FILE,
         help="config file to parse the data",
     )
+    parser.add_argument(
+        SKIP_SPLITS_CHECK_CLI_ARGUMENT,
+        dest=SKIP_SPLITS_CHECK,
+        action="store_true",
+        default=False,
+        help="whether to skip automatic data splits check",
+    )
     return vars(parser.parse_args())
 
 
@@ -34,5 +42,8 @@ def main():
     sys.path.append(os.getcwd())
 
     args = get_args_dict()
+
     options = yaml.load(open(args[CONFIG_FILE], "r"), Loader=yaml.FullLoader)
+    options.update({SKIP_SPLITS_CHECK: args[SKIP_SPLITS_CHECK]})
+
     preprocess_data(options)